+++ src/frames/swap/index.ts
@@ -5,6 +5,7 @@ import { getFont } from '../fonts'
 import { transaction } from './transaction'
 import { finishScreen } from './finish'
 import { tradeScreen } from './trade'
+import { tokenSelectionScreen } from './token-selection'
 
 type FrogOptions = {
   Bindings: { ZEROX_API_KEY?: string }
@@ -19,6 +20,7 @@ export const app = new Frog<FrogOptions>({
 })
 
 app.frame('/', homeScreen)
+app.frame('/token-selection', tokenSelectionScreen)
 app.frame('/:network/:token', tradeScreen)
 app.frame('/finish', finishScreen)
 app.transaction('/tx', transaction)
